--- frontend/apps/app/features/schemas/components/Chatbot/index.ts
@@ -1,5 +1,0 @@
-export * from './AskAiIcon'
-export * from './ChatbotButton'
-export * from './ChatbotDrawer'
-export * from './ChatInput'
-export * from './ChatMessage'
