+++ frontend/packages/erd-core/src/features/erd/components/ERDRenderer/ERDRenderer.tsx
@@ -12,7 +12,6 @@ import {
 import { ReactFlowProvider } from '@xyflow/react'
 import {
   type ComponentProps,
-  type ComponentType,
   type FC,
   createRef,
   useCallback,
@@ -32,13 +31,7 @@ import { LeftPane } from './LeftPane'
 import { RelationshipEdgeParticleMarker } from './RelationshipEdgeParticleMarker'
 import { TableDetailDrawer, TableDetailDrawerRoot } from './TableDetailDrawer'
 import { Toolbar } from './Toolbar'
-
-import type { Schema } from '@liam-hq/db-structure'
-
-interface ChatbotButtonProps {
-  schemaData: Schema
-  tableGroups?: Record<string, TableGroup>
-}
+import type { ChatbotButtonComponentType } from './Toolbar/types'
 
 type Props = {
   defaultSidebarOpen?: boolean | undefined
@@ -47,7 +40,7 @@ type Props = {
   withAppBar?: boolean
   tableGroups?: Record<string, TableGroup>
   onAddTableGroup?: ((params: TableGroup) => void) | undefined
-  ChatbotButtonComponent?: ComponentType<ChatbotButtonProps>
+  ChatbotButtonComponent?: ChatbotButtonComponentType
 }
 
 const SIDEBAR_COOKIE_NAME = 'sidebar:state'
