--- frontend/packages/erd-core/src/types.ts
@@ -1,5 +1,0 @@
-export interface TableGroupData {
-  name?: string
-  tables?: string[]
-  comment?: string | null
-}
